+++ front/src/app/presentation/[id]/page.tsx
@@ -1,45 +1,68 @@
-'use client'
-
-import { useEffect, useState } from 'react';
+'use client';
+import { Navigation, Pagination } from 'swiper/modules';
+import { Swiper, SwiperSlide } from 'swiper/react';
+import { Mediapipe } from '../../components/Detection/Mediapipe';
+import { Effects } from '../../components/Effects/Effects';
+import { Sounds } from '../../components/Sounds/Sounds';
+import * as pdfjsLib from 'pdfjs-dist';
+import 'pdfjs-dist/build/pdf.worker.min.mjs';
+import { usePdfSlider } from '../usePdfSlider';
+import { useSlideNavigation } from '../useSlideNavigation';
 import { useParams } from 'next/navigation';
 
-export default function PdfViewer() {
-    const [pdfUrl, setPdfUrl] = useState('');
-    const params = useParams();
-    const id = params.id;
+pdfjsLib.GlobalWorkerOptions.workerSrc = '/pdf.worker.min.mjs';
 
-    useEffect(() => {
-        const fetchPdf = async () => {
-            try {
-                // GinサーバーのURLに合わせて調整
-                const response = await fetch(`http://localhost:8080/api/pdf/${id}`, {
-                    method: 'GET',
-                    credentials: 'include',
-                    headers: {
-                        'Content-Type': 'application/json',
-                    }
-                });
-                if (!response.ok) {
-                    throw new Error('Network response was not ok');
-                }
-                const blob = await response.blob();
-                const url = URL.createObjectURL(blob);
-                setPdfUrl(url);
-            } catch (error) {
-                console.error('Error fetching PDF:', error);
-            }
-        };
+const slideSettings = {
+  0: {
+    slidesPerView: 1,
+    spaceBetween: 10,
+  },
+  1024: {
+    slidesPerView: 1,
+    spaceBetween: 10,
+  },
+};
 
-        fetchPdf();
-    }, []);
+const BasicSlider: React.FC = () => {
+  const params = useParams();
+  const pdfid = params.id as string;
+  
+  const { swiperInstanceRef, prevSlide, nextSlide } = useSlideNavigation();
+  const [ images ] = usePdfSlider(swiperInstanceRef, pdfid);
 
-    return (
-        <div>
-            {pdfUrl ? (
-                <iframe src={pdfUrl} width="100%" height="100%" style={{ minHeight: '80vh' }}></iframe>
-            ) : (
-                <p>Loading...</p>
-            )}
-        </div>
-    );
-}
+  return (
+    <div>
+      <div className="static">
+        <Swiper
+          modules={[Navigation, Pagination]}
+          breakpoints={slideSettings}
+          slidesPerView="auto"
+          centeredSlides
+          loop
+          speed={1000}
+          pagination={{ clickable: true }}
+          initialSlide={0}  // 初期表示で1ページ目を指定
+          className="h-dvh"
+          onSwiper={(swiper) => { swiperInstanceRef.current = swiper; }}
+        >
+          {images.map((src, index) => (
+            <SwiperSlide key={index}>
+              <img
+                src={src}
+                alt={`Slider Image ${index + 1}`}
+                className="h-full object-cover w-full"
+              />
+            </SwiperSlide>
+          ))}
+        </Swiper>
+      </div>
+      <div className="absolute top-0 right-0 z-10" >
+        <Mediapipe nextSlide={nextSlide} prevSlide={prevSlide} />
+      </div>
+      <Effects/>
+      <Sounds/>
+    </div>
+  );
+};
+
+export default BasicSlider;
